# Security: Optimize CodeQL configuration with battle-tested config + ReDoS fix #777