… .github/workflows/claude-comment-gate.yml
… .github/workflows/claude-issue-review.yml
… .github/workflows/claude-pr-review-labeled.yml
… .github/workflows/codeql.yml
… codeql-config.yml
… docs/logging/secure-logging.md
… test/setup.ts
… test/unit/utils/secure-logger-edge-cases.test.ts
… test/unit/utils/secure-logger.test.ts