@kesslerio 2025-09-29
Closed: Replaced by clean branch (feature/issue-777-codeql-optimization) to resolve merge conflicts